Try to address non download capable watches
Some users are installing the app on a non download capable watch,
then being disappointed when they cannot download.

Unfortunately we do not have enough space to explain fully on the watch
UI, and the space above the fold on the app page is quite limited.

Best compromise appears to be to upload a second copy of the app as a 'lite'
version, and restrict the original version to download capable watches